[reorg] Move NER script out of bert folder  (#1090)
* first commit

* rename

* update doc

* fix a bug

* fix import

* add test

* Update py3-master.yml

* Update py3.yml

* Fix test invocation

* Update test_scripts.py

* Update test_scripts.py

* Update test_scripts.py

* fix batch size in the test

* Update model_zoo.rst
