+++ scripts/sequence_labeling/dataset_sample/test_sample.txt
@@ -1,0 +1,17 @@
+-DOCSTART- -X- -X- O
+
+SOCCER NN I-NP O
+- : O O
+JAPAN NNP I-NP I-LOC
+GET VB I-VP O
+LUCKY NNP I-NP O
+WIN NNP I-NP O
+, , O O
+CHINA NNP I-NP I-PER
+IN IN I-PP O
+SURPRISE DT I-NP O
+DEFEAT NN I-NP O
+. . O O
+
+Nadim NNP I-NP I-PER
+Ladki NNP I-NP I-PER
