+++ scripts/sequence_labeling/predict_ner.py
@@ -24,8 +24,8 @@ import os
 import mxnet as mx
 from ner_utils import get_bert_model, get_context
 from ner_utils import load_metadata
-from data.ner import BERTTaggingDataset, convert_arrays_to_text
-from model.ner import BERTTagger
+from data import BERTTaggingDataset, convert_arrays_to_text
+from model import BERTTagger
 
 # TODO(bikestra): Currently, our evaluation is dependent on this package.
 # Figure out whether to take actual dependency on it.
