+++ scripts/tests/test_scripts.py
@@ -353,3 +353,20 @@ def test_xlnet_finetune_glue(dataset):
     process = subprocess.check_call([sys.executable, './scripts/language_model/run_glue.py']
                                     + arguments)
     time.sleep(5)
+
+@pytest.mark.serial
+@pytest.mark.gpu
+@pytest.mark.remote_required
+@pytest.mark.integration
+def test_bert_ner():
+    folder = './scripts/sequence_labeling'
+    arguments = ['--train-path', folder + '/dataset_sample/train_sample.txt',
+                 '--dev-path', folder + '/dataset_sample/validation_sample.txt',
+                 '--test-path', folder + '/dataset_sample/test_sample.txt',
+                 '--gpu', '0', '--learning-rate', '1e-5',
+                 '--warmup-ratio', '0', '--batch-size', '1',
+                 '--num-epochs', '1', '--bert-model', 'bert_24_1024_16',
+                 '--save-checkpoint-prefix', './test_bert_ner']
+    script = folder + '/finetune_bert.py'
+    process = subprocess.check_call([sys.executable, script] + arguments)
+    time.sleep(5)
